+++ app/src/FreeDeck.h
@@ -1,4 +1,5 @@
 #include <Arduino.h>
+
 extern int currentPage;
 extern int pageCount;
 int getCurrentPage();
@@ -7,7 +8,7 @@ void setMuxAddress(int address);
 void displayImage(int16_t imageNumber);
 void loadPage(int16_t pageIndex);
 void executeButtonConfig(uint8_t buttonIndex, uint8_t buttonUp,
-						 uint8_t secondary);
+                         uint8_t secondary);
 void checkButtonState(uint8_t buttonIndex);
 void initAllDisplays();
 void setGlobalContrast();
